Resolve issue with php strict type (#1110)
* Resolve issue with php strict type

related to the issue https://github.com/laravel/horizon/issues/1109

* Update RedisMetricsRepository.php

+++ src/Repositories/RedisMetricsRepository.php
@@ -331,8 +331,8 @@ class RedisMetricsRepository implements MetricsRepository
      */
     protected function minutesSinceLastSnapshot()
     {
-        $lastSnapshotAt = $this->connection()->get('last_snapshot_at')
-                    ?: $this->storeSnapshotTimestamp();
+        $lastSnapshotAt = (int) ($this->connection()->get('last_snapshot_at')
+                                    ?: $this->storeSnapshotTimestamp());
 
         return max(
             (CarbonImmutable::now()->getTimestamp() - $lastSnapshotAt) / 60, 1
